fix: profile handling

+++ src/components/profileslist.ts
@@ -3,14 +3,16 @@ import ProfileService from "../lib/profile";
 
 export default class ProfilesList implements m.Component {
   deleteProfile(profileName: string) {
-    ProfileService.deleteProfile(profileName);
-    m.redraw();
+    if (window.confirm(`Are you sure you want to delete the profile: ${profileName}?`)) {
+      ProfileService.deleteProfile(profileName);
+      m.redraw();
+    }
   }
 
   view() {
     const profileNames = ProfileService.getProfileIds();
     return m("div", profileNames.map(name => {
-      return m(".box.profile", [
+      return m(".box.profile", { style: { marginBottom: '15px' } }, [
         m(".content", [
           m("p.title.is-5", name),
           m("div", [
